refactor 24.10.07 syb calculator에서 연산 부분 분리
